fixes, float type in migrations

+++ database/migrations/2021_03_14_172221_create_birds_table.php
@@ -17,7 +17,7 @@ class CreateBirdsTable extends Migration
             $table->id();
             $table->string('name')->default('');
             $table->integer('price')->default(0);
-            $table->integer('fertility')->default(0);
+            $table->float('fertility')->default(0);
             $table->text('description')->nullable();
             $table->text('image')->nullable();
             $table->timestamps();
