+++ database/migrations/2021_04_17_122844_alter_table_birds_add_egs_price.php
@@ -14,7 +14,7 @@ class AlterTableBirdsAddEgsPrice extends Migration
     public function up()
     {
         Schema::table('birds', function (Blueprint $table) {
-            $table->integer('egg_price')->default(1)->after('fertility');
+            $table->float('egg_price')->default(1)->after('fertility');
         });
     }
